refactor: structure file move all controllers and services file into modules file
